Untested fix of mix up between hue keys and chroma values used in lerp call
Untested. Previous code seemed to mix up hue key and chroma value. With the previous code the lerp call would return (hue/(high-low))*(high-low) which would be hue but with potential massive rounding errors.

This might fix some of the strange jumping selections when switching between modes and then adjusting sliders. This changed lerp could still possibly give a max chroma that is a tiny bit too large, I think, if the max chroma curve is convex between hue_low and hue_high. But this would probably be rare since the values in the chroma tree are approximate, and so slightly below anyway.

Another thing that might become a problem is hues higher than 178/180*pi get aproximate max chroma of hue 178/180*pi. If actual max chromas were decreasing for that small slice of hues that'd be a likely place to pick a max chroma too high. Not sure how important this is. Don't understand most of what is going on.